+++ lib/cocoapods-core/specification/linter.rb
@@ -351,7 +351,7 @@ module Pod
       # Performs validations related to the `deprecated_in_favor_of` attribute.
       #
       def _validate_deprecated_in_favor_of(d)
-        if d == Specification.root_name(d)
+        if spec.root.name == Specification.root_name(d)
           results.add_error('deprecated_in_favor_of', 'a spec cannot be ' \
             'deprecated in favor of itself')
         end